Implemented separate function to convert the array of errors into a nested dictionary of errors from the validator
