+++ software/core/Infrastructure/backend/dataCorrector/corrector.js
@@ -1,23 +1,55 @@
 const fs = require("fs");
 
-var commonErrors = fs.readFileSync("./commonErrors.json");
-var commonErrorsParsed = JSON.parse(commonErrors);
-var commonErrorkeyArray = Object.keys(commonErrorsParsed);
+// var commonErrors = fs.readFileSync("./commonErrors.json");
+// var commonErrorsParsed = JSON.parse(commonErrors);
+// var commonErrorkeyArray = Object.keys(commonErrorsParsed);
 
-function corrector(json_odriod_data, error_obj) {
+function corrector(json_odriod_data, error_array) {
 
-    for (const [dataArray, sensor] of Object.entries(error_obj)) {
-        for (const [sensor_key, error] of Object.entries(sensor)){
-            for (const [error_key, value] of Object.entries(error)){
-                // console.log(error_key)
-                if (error_key == "speedNotFoundError") insert_speed();
 
-            }
-        }
-    }
+
+    // for (const [dataArray, sensor] of Object.entries(error_obj)) {
+    //     for (const [sensor_key, error] of Object.entries(sensor)){
+    //         for (const [error_key, value] of Object.entries(error)){
+    //             // console.log(error_key)
+    //             // if (error_key == "speedNotFoundError") insert_speed();
+
+    //         }
+    //     }
+    // }
 
 }
 
+var error_array =
+[
+    {
+      errorId: 1,
+      dataArrayName: 'speedDataArray',
+      entity: 1,
+      error: 'negativeValueError'
+    },
+    {
+      errorId: 2,
+      dataArrayName: 'speedDataArray',
+      entity: 1,
+      error: 'nameNotFoundError'
+    },
+    {
+      errorId: 3,
+      dataArrayName: 'temperatureDataArray',
+      entity: 1,
+      thresholdValue: 1000,
+      error: 'nameNotFoundError'
+    }
+]
+
+error_array.forEach(elem => {
+    // console.log(elem.error);
+});
+
+/*
+------------------------------------------------------
+*/
 
 var temp_error = {
     "brakeDataArray": {
@@ -37,6 +69,9 @@ var temp_error = {
 // if (temp_error.brakeDataArray.brakes.brakesNotFoundError) console.log("true");
 // else console.log("false"); 
 
+/*
+------------------------------------------------------------
+*/
 
 var obj = {
     "speed": [
@@ -54,6 +89,12 @@ var hello = "hello";
 obj["speed"][0][temp_string] = hello;
 
 // console.log(JSON.stringify(obj, null, 4));
+
+
+
+
+
+
 
 var test_data = fs.readFileSync("./test_data.json");
 var test_errors = fs.readFileSync("./test_error.json");
