+++ software/core/Infrastructure/backend/index.js
@@ -7,6 +7,7 @@ const encapsulate = require("./Encapsulate/index");
 const broadcast = require("./utilities/broadcast");
 const mock = require("./utilities/mock");
 const validate = require("./validator/validate");
+const validate = require("./validator/createErrorObj");
 const corrector = require("./dataCorrector/corrector");
 
 const wss = new WebSocket.Server({ port: 8080 });
@@ -38,14 +39,18 @@ wss.on("connection", function connection(ws) {
 					//Todo: add encapsulation for the massage to go to the odroid
 					broadcast(wss, message, "odroid");
 				} else {
-					var errorObj = {};
+					var error_array = [];
+					var error_obj = {};
 
 					// validate incoming data from the pod
-					var errorObj = validate(data);
+					var error_array = validate(data);
 		
+					// get errors as a nested dictionary
+					error_obj = createErrorObj(error_array);
+
 					// corrected json data
 					// var correctedData;
-					// correctedData = corrector(data, errorObj);
+					// correctedData = corrector(data, error_array);
 
 					// Divide incoming data into multiple components
 					const Digestor = digest(correctedData);
